backend/requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
dj-rest-auth==6.0.0
Django==5.1a1
django-allauth==0.63.3
django-cors-headers==4.4.0
djangorestframework==3.15.1
execnet==2.1.1
idna==3.7
iniconfig==2.0.0
munkres==1.1.4
networkx==3.3
numpy==2.0.0
packaging==24.1
pluggy==1.5.0
pytest==8.2.2
pytest-django==4.8.0
pytest-xdist==3.6.1
requests==2.32.3
scipy==1.13.1
sqlparse==0.5.0
urllib3==2.2.2
